scripts/requirements.txt
flatbuffers 
tflite
tensorflow
Pillow
serial
numpy
